# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Simulated login. No JWT; validates email/password and returns session object.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/me/{userId}": {
      "get": {
        "operationId": "AuthController_me",
        "summary": "Get current user profile by user id",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/assignments": {
      "get": {
        "operationId": "AssignmentsController_all",
        "summary": "List assignments",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Assignments"
        ]
      },
      "post": {
        "operationId": "AssignmentsController_create",
        "summary": "Create assignment",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssignmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Assignments"
        ]
      }
    },
    "/api/assignments/learner/{learnerId}": {
      "get": {
        "operationId": "AssignmentsController_byLearner",
        "summary": "List learner assignments",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          },
          {
            "name": "learnerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Assignments"
        ]
      }
    },
    "/api/assignments/{id}": {
      "get": {
        "operationId": "AssignmentsController_one",
        "summary": "Get one assignment",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Assignments"
        ]
      },
      "patch": {
        "operationId": "AssignmentsController_update",
        "summary": "Update assignment",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAssignmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Assignments"
        ]
      },
      "delete": {
        "operationId": "AssignmentsController_remove",
        "summary": "Delete assignment",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Assignments"
        ]
      }
    },
    "/api/assignments/{id}/submit": {
      "post": {
        "operationId": "AssignmentsController_submit",
        "summary": "Submit assignment",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAssignmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Assignments"
        ]
      }
    },
    "/api/assignments/{id}/approve/{learnerId}": {
      "post": {
        "operationId": "AssignmentsController_approve",
        "summary": "Approve learner assignment completion",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "learnerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Assignments"
        ]
      }
    },
...
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Public user registration compatibility endpoint. Creates a user without JWT.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "IndAI Review-4 Paper Trading API",
    "description": "NestJS modular backend with in-memory arrays, RBAC through x-role header, DTO validation, Swagger and minute-stock data endpoints.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "learner@indai.com",
            "description": "Required. Frontend allows Gmail or email ending with .in only."
          },
          "password": {
            "type": "string",
            "example": "Learn@1234",
            "description": "Required password."
          },
          "role": {
            "type": "string",
            "enum": [
              "superuser",
              "admin",
              "instructor",
              "provider",
              "learner",
              "SUPER_USER",
              "ADMIN",
              "INSTRUCTOR",
              "COURSE_PROVIDER",
              "LEARNER"
            ],
            "example": "LEARNER",
            "description": "Optional selected role from frontend login/register."
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "New",
            "minLength": 3,
            "description": "Required. Alphabets and spaces only."
          },
          "lastName": {
            "type": "string",
            "example": "Learner",
            "minLength": 3,
            "description": "Required. Alphabets and spaces only."
          },
          "email": {
            "type": "string",
            "example": "new@gmail.com",
            "description": "Required. Gmail or .in email only."
          },
          "password": {
            "type": "string",
            "example": "Test@1234",
            "minLength": 8,
            "description": "Required. Min 8 chars, uppercase, lowercase, number, and one of !@#$%^&*."
          },
          "confirmPassword": {
            "type": "string",
            "example": "Test@1234",
            "description": "Optional compatibility field used by frontend/Swagger. If provided, it must match password."
          },
          "role": {
            "type": "string",
            "enum": [
              "superuser",
              "admin",
              "instructor",
              "provider",
              "learner",
              "SUPER_USER",
              "ADMIN",
              "INSTRUCTOR",
              "COURSE_PROVIDER",
              "LEARNER"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "pending",
              "suspended",
              "disabled"
            ],
            "example": "active"
          },
          "phone": {
            "type": "string",
            "example": "9876543210",
            "description": "Optional but if provided must be 10 digits starting with 6/7/8/9."
          },
          "dateOfBirth": {
            "type": "string",
            "example": "2007-05-09",
            "description": "Required by registration frontend. Learner must be at least 10 years old."
          },
          "institution": {
            "type": "string",
            "example": "IIIT Sri City",
            "minLength": 3
          },
          "studentId": {
            "type": "string",
            "example": "CSE2026",
            "minLength": 3
          },
          "grade": {
            "type": "string",
            "example": "BTech 2nd Year",
            "minLength": 3
          },
          "major": {
            "type": "string",
            "example": "Computer Science",
            "minLength": 3
          },
          "organization": {
            "type": "string",
            "example": "IndAI Academy",
            "minLength": 3
          },
          "expertise": {
            "type": "string",
            "example": "Technical Analysis",
            "minLength": 3
          },
          "yearsExp": {
            "type": "string",
            "example": "3-5"
          },
          "website": {
            "type": "string",
            "example": "https://example.com"
          },
          "contentType": {
            "type": "string",
            "example": "Video Courses"
          },
          "authCode": {
            "type": "string",
            "example": "ADMIN-2026"
          },
          "accessLevel": {
            "type": "string",
            "example": "Full"
          },
          "tradingExperience": {
            "type": "string",
            "enum": [
              "Beginner",
              "Intermediate",
              "Advanced",
              "BEGINNER",
              "INTERMEDIATE",
              "ADVANCED",
              "0-1",
              "1-3",
              "3-5",
              "5+"
            ]
          },
          "riskTolerance": {
            "type": "string",
            "enum": [
              "Low",
              "Medium",
              "High",
              "Conservative",
              "Moderate",
              "Aggressive",
              "LOW",
              "MEDIUM",
              "HIGH"
            ]
          },
          "startingBalance": {
            "type": "number",
            "example": 100000
          },
          "tradingLimit": {
            "type": "number",
            "example": 150000
          },
          "instructorId": {
            "type": "string",
            "example": "u3"
          },
          "skillPoints": {
            "type": "number",
            "example": 50
          },
          "portfolioValue": {
            "type": "number",
            "example": 100000
          },
          "virtualBalance": {
            "type": "number",
            "example": 100000
          },
          "experience": {
            "type": "string",
            "example": "Beginner"
          },
          "goals": {
            "type": "string",
            "example": "Learn paper trading safely"
          },
          "studentIds": {
            "example": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "yearsOfExperience": {
            "type": "number",
            "example": 5,
            "description": "Compatibility alias for yearsExp used by some Swagger/frontend payloads."
          },
          "experienceLevel": {
            "type": "string",
            "enum": [
              "Beginner",
              "Intermediate",
              "Advanced",
              "BEGINNER",
              "INTERMEDIATE",
              "ADVANCED",
              "0-1",
              "1-3",
              "3-5",
              "5+"
            ],
            "description": "Compatibility alias for experience/tradingExperience."
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "password",
          "role"
        ]
      },
...
      "CreateAssignmentDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Trading Plan",
            "minLength": 3
          },
          "description": {
            "type": "string",
            "example": "Submit a paper trade plan.",
            "minLength": 3
          },
          "instructorId": {
            "type": "string",
            "example": "u3"
          },
          "studentIds": {
            "example": [
              "u6"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dueDate": {
            "type": "string",
            "example": "2026-05-20",
            "description": "Must be a future date within 365 days."
          },
          "skillPoints": {
            "type": "number",
            "example": 50,
            "minimum": 1,
            "maximum": 200
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "Easy",
              "Medium",
              "Hard",
              "BEGINNER",
              "INTERMEDIATE",
              "ADVANCED",
              "Beginner",
              "Intermediate",
              "Advanced"
            ],
            "example": "Medium"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "completed",
              "pending",
              "in_progress",
              "submitted",
              "approved"
            ],
            "example": "active"
          },
          "type": {
            "type": "string",
            "example": "Trading Challenge"
          }
        },
        "required": [
          "title",
          "description",
          "instructorId",
          "dueDate",
          "skillPoints"
        ]
      },
      "UpdateAssignmentDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 3
          },
          "description": {
            "type": "string",
            "minLength": 3
          },
          "instructorId": {
            "type": "string"
          },
          "studentIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dueDate": {
            "type": "string"
          },
          "skillPoints": {
            "type": "number",
            "minimum": 1,
            "maximum": 200
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "Easy",
              "Medium",
              "Hard",
              "BEGINNER",
              "INTERMEDIATE",
              "ADVANCED",
              "Beginner",
              "Intermediate",
              "Advanced"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "completed",
              "pending",
              "in_progress",
              "submitted",
              "approved"
            ]
          },
          "type": {
            "type": "string"
          }
        }
      },
      "SubmitAssignmentDto": {
        "type": "object",
        "properties": {
          "learnerId": {
            "type": "string",
            "example": "u6"
          },
          "submissionText": {
            "type": "string",
            "example": "My solution text"
          }
        },
        "required": [
          "learnerId"
        ]
      },
...
    }
  }
}